FinalProject/raspberry/scripts/requirements.txt
bme280pi==1.1.2
RPi.GPIO==0.7.1
smbus==1.1.post2